+++ src/index.ts
@@ -3,7 +3,7 @@ export * from "./interfaces";
 
 export { runTaskList, runTaskListOnce } from "./main";
 export { run, runOnce, runMigrations } from "./runner";
-export { WorkerUtils, addJob } from "./worker_utils";
+export { makeWorkerUtils, addJob } from "./workerUtils";
 export { Logger, LogFunctionFactory, consoleLogFactory } from "./logger";
 
 export { getTasks };
